% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arscpp.R
\name{mplnposterior}
\alias{mplnposterior}
\title{Sample from a multivariate Poisson-Log-Normal posterior}
\usage{
mplnposterior(y, mu, H)
}
\arguments{
\item{y}{A numeric vector of counts of size N.}

\item{mu}{The prior mean vector of size N.}

\item{H}{A sparse precision matrix of size NxN, preferably in 'dsCMatrix' (sparse column) format.}
}
\value{
A \code{MPLNPosterior} object.
}
\description{
\code{mplnposterior} returns an object for sampling from the MPLN posterior
using an ARS-in-Gibbs sampler.
}
\details{
The returned object samples from

\deqn{x \sim p(x | y, \mu, \Sigma) \propto \prod_i [dPois(y_i; exp(x_i))] \times dMVNorm(x; \mu, \Sigma)}
}
\examples{
N <- 100
mu <- rep(0, N)
sd <- rep(2, N)
x <- rnorm(N, mu, sd)
y <- rpois(N, exp(x))
H <- as(diag(1/(sd^2)), 'dsCMatrix')
mpln <- mplnposterior(y, mu, H)
smp <- mpln$sample(100, x) # x are starting values

}

