% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arscpp.R
\name{plnposterior}
\alias{plnposterior}
\title{Sample from a Poisson-Log-Normal posterior}
\usage{
plnposterior(y, mu, sd)
}
\arguments{
\item{y}{A numeric vector of counts.}

\item{mu}{A numeric vector of prior means (must be of same length as \code{y}).}

\item{sd}{A numeric vector of prior sds (must be of same length as \code{y}).}
}
\value{
A \code{PLNPosterior} object.
}
\description{
\code{plnposterior} returns an object for sampling from the PLN posterior.
}
\details{
The returned object samples from

\deqn{x \sim p(x | y, \mu, sd) \propto dPois(y; exp(x)) \times dNorm(x; \mu, sd)}
}
\examples{
N <- 100
mu <- rep(0, N)
sd <- rep(1, N)
x <- rnorm(N, mu, sd)
y <- rpois(N, exp(x))
pln <- plnposterior(y, mu, sd)
smp <- pln$sample(100)

}

